requirements.txt
absl-py==0.8.0
asn1crypto==0.24.0
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
blinker==1.4
brotlipy==0.7.0
catalyst==20.7
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
click==6.7
colorama==0.4.1
cryptography==2.3.1
cycler==0.10.0
Cython==0.29.18
decorator==4.4.0
defusedxml==0.6.0
deprecation==2.1.0
efficientnet==1.1.0
entrypoints==0.3
fast-rcnn==0.0.0
future==0.18.2
gast==0.3.2
gitdb==4.0.5
GitPython==3.1.7
google-pasta==0.1.7
graphviz==0.13.2
grpcio==1.23.0
h11==0.7.0
h2==3.1.1
h5py==2.9.0
hpack==3.0.0
hyperframe==5.2.0
idna==2.8
imageio==2.5.0
imgaug==0.4.0
importlib-metadata==1.3.0
imutils==0.5.3
ipykernel==5.1.3
ipython==7.10.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.1
jsons==1.0.0
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
kaitaistruct==0.8
Keras==2.2.4
keras-adabound==0.5.0
Keras-Applications==1.0.7
keras-lookahead==0.9.0
Keras-Preprocessing==1.1.0
keras-radam==0.15.0
kiwisolver==1.1.0
ldap3==2.5.2
loguru==0.3.2
lxml==4.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
mitmproxy==4.0.4
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.3
nms==0.0.0
notebook==6.0.2
numexpr==2.7.1
numpy==1.19.0
opencv-contrib-python==4.1.2.30
opencv-python==3.4.4.19
opencv-python-headless==4.1.1.26
orca==1.5.3
packaging==20.4
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
passlib==1.7.1
pickleshare==0.7.5
Pillow==6.0.0
plotly==4.5.4
prettytable==0.7.2
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.12.2
psutil==5.7.0
py-fast-rcnn-gpu==0.0.0
pyasn1==0.4.7
pycocotools @ git+https://github.com/philferriere/cocoapi.git@2929bd2ef6b451054755dfd7ceb09278f935f7ad#subdirectory=PythonAPI
pycparser==2.19
pydivert==2.1.0
pydot==1.2.3
Pygments==2.5.2
pynput==1.6.8
pyOpenSSL==18.0.0
pyparsing==2.4.6
pyperclip==1.6.5
PyQt5==5.14.2
PyQt5-sip==12.7.2
pyrsistent==0.15.6
python-dateutil==2.8.1
pythonping==1.0.5
pytorch-ranger==0.1.1
pytz==2019.3
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.1.2
pyzmq==18.1.1
requests==2.23.0
retrying==1.3.3
ruamel.yaml==0.15.100
scikit-image==0.17.2
scikit-learn==0.22
scipy==1.3.0
seaborn==0.9.0
Send2Trash==1.5.0
Shapely==1.6.4.post2
six==1.13.0
sklearn==0.0
smmap==3.0.4
sortedcontainers==2.0.5
tables==3.6.1
tabulate==0.8.5
tensorboard==1.14.0
tensorboardX==2.0
tensorflow-determinism==0.3.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tifffile==2020.7.24
toolz==0.10.0
torch==1.5.1+cu101
torch-optimizer==0.0.1a13
torchvision==0.6.1+cu101
tornado==5.1.1
tqdm==4.36.1
traitlets==4.3.3
typish==1.2.0
urllib3==1.25.8
urwid==2.0.1
wcwidth==0.1.7
webcolors==1.11.1
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
win32-setctime==1.0.1
wrapt==1.11.2
wsproto==0.11.0
xgboost==1.1.1
XlsxWriter==1.2.6
xlwt==1.3.0
zipp==0.6.0
